+++ js/data.js
@@ -1,0 +1,230 @@
+import {getRandomInt, getRandomArrayElem} from './util.js';
+
+let id = 0;
+
+const MAX_ID = 25;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 200;
+
+const getId = () => ++id;
+
+const getUrl = () => `photos/${id}.jpg`;
+
+// комменты с сайта
+// https://notagram.ru/200-krutyh-podpisej-dlya-vashih-foto-v-instagram/
+const description = [
+  'Если смогу, я сделаю это. Конец истории.',
+  'Смейтесь как только умеете, любите столько, сколько живете.',
+  'Помните: вы единственный человек, который может наполнить ваш мир солнечным светом.',
+  'Я полностью уверена, что я — диснеевская принцесса, которую еще не придумали.',
+  'Не позволяйте кому-то затушить ваши искры только потому, что их свет сияет в чьих-то глазах.',
+  'Делайте в вашей жизни то, что меньше заставляет вас смотреть в свой телефон.',
+  'Улыбка — единственный тренд в моде, который актуален всегда.',
+  'Никогда не ищите свое счастье там, где вы его однажды потеряли.',
+  'Жизнь похожа на фотокамеру: вам просто нужно смотреть на нее с улыбкой.',
+  'Моя жизнь меняется, потому что меняю ее я.',
+  'Всегда начинайте свой день с хороших людей и кофе.',
+  'Ни о чем не беспокойтесь. Потому что все лучшие умы на работе.',
+  'Жизнь — это всего лишь серия крошечных чудес, поэтому обратите внимание на них.',
+  'Живите во всех тех моментах, которые вы не можете выразить словами.',
+  'Не ждите идеального момента. Берите каждый момент и делайте его идеальным.',
+  'Признай это. Без меня, твоя жизнь была бы действительно скучной.',
+  'Будьте счастливы в этот момент, потому что этот момент — и есть ваша жизнь.',
+  'Я пыталась заниматься йогой, но в позе лотоса уснула.',
+  'Я, возможно, никогда не буду лучшей, но я стараюсь быть лучшей твоей.',
+  'Если вам никто не улыбнулся утром, я подарю вам одну из своих.',
+  'Никогда не позволяйте никому скучать.',
+  'Все только начинает становиться действительно хорошим.',
+  'Я опять съела сладкое. А все потому, что каждую секунду в мире 200 человек празднуют свой день рождения!',
+  'Мечтайте. Поверьте, в это. Добейтесь этого.',
+  'Утром, только одна хорошая мысль меняет смысл целого дня.',
+  'Надейтесь на лучшее, но не ждите этого. Смотрите вперед в будущее, но никогда не ждите, пока это произойдет.',
+  'Любите меня, от этого я буду сиять еще ярче.',
+  'Я точно знаю, кто я, и я чертовски горжусь этим.',
+  'Никогда не бывает хорошего результата, если у вас негативное отношение к этому.',
+  'Просто ешьте, живите и любите. Затем с утра повторяйте каждый день.',
+  'Лучший способ предсказать свое будущее — это активно его создавать.',
+  'Я никогда ничего не теряю. Я либо побеждаю, либо узнаю что-то новое.',
+  'Мне нравится и улыбаюсь!',
+  'Я не толстый. Меня просто легче увидеть, чем всех остальных.',
+  'Улыбайтесь, потому что это делает всех нас красивыми.',
+  'Это просто моя жизнь в моем неповторимом стиле.',
+  'Я прекрасна, как Бог создал меня.',
+  'Не заботьтесь ни о чем, больше улыбайтесь.',
+  'Я не ленивый. Просто у меня нет мотивации.',
+  'Независимо от того, что вы делаете в жизни, убедитесь, что это то, что делает вас счастливыми.',
+  'Все, кем я хочу быть, это тот, кем я был.',
+  'Если у вас есть глаза, взгляните на меня сейчас!',
+  'Лучшая вещь, которую женщина может носить, чтобы выглядеть потрясающе, — это ее уверенность.',
+  'Воскресенье — еще один способ сказать: «Какой чудесный день!»',
+  'Предупреждение: на этом фото вы можете влюбиться в меня.',
+  'Лучшее еще впереди.',
+  'Жизнь похожа на фотографию. Мы развиваемся только из негативов.',
+  'Предупреждение: мой рот работает быстрее, чем мой мозг.',
+  'Жизнь слишком коротка, чтобы тратить время на то, чтобы найти все ответы. Вместо этого, наслаждайтесь всеми вопросами!',
+  'Мне нравится думать, что я на диете. Как же это ужасно!',
+  'Окружайте себя настоящими друзьями, и вы будете действительно счастливы.',
+  'Всегда есть дикая сторона, которая скрывается за самым невинным лицом.',
+  'Выходные, пожалуйста, не оставляйте меня сейчас!',
+  'В простоте есть удивительная красота.',
+  'Это моя жизнь, и мне так повезло ее жить.',
+  'Я — это мы.',
+  'Никогда не бойтесь. Просто постарайтесь быть собой.',
+  'Бог должен быть невероятно креативным. Что я имею в виду? Просто посмотри на меня!',
+  'Да, я сумасшедшая. Быть нормальной — слишком скучно.',
+  'В любой ситуации всегда улыбайтесь.',
+  'Я просто увидел самого умного человека в мире, когда смотрел в зеркало.',
+  'Всегда помните: вы живете только один раз.',
+  'Дорога к успеху строится постоянно.',
+  'Правило жизни: никогда не проверяйте глубину воды обеими ногами.',
+  'Всегда будьте лучшим вариантом для себя.',
+  'Я не шопоголик. Я просто вношу свою лепту, чтобы помочь нашей экономике.',
+  'Мне не нужно одобрять себя.',
+  'В море может и водится миллион рыб, но я единственная русалка.',
+  'Доброе утро, всем! Теперь давайте начнем стресс!',
+  'Лучшая подготовка к завтрашнему дню — это сделать все возможное сегодня.',
+  'Жить — это так здорово.',
+  'Помните: не все, кто ушел, уходят.',
+  'Я устал сегодня. Пожалуйста, дайте мне новый день.',
+  'Никто никогда не совершенен. Вот почему изобрели корректор и ластик.',
+  'Как бы вы ни чувствовали себя сегодня — вставайте, одевайтесь, покажите всем, что никогда не сдаетесь.',
+  'Шоппинг — это особый вид искусства. Уважайте художника, пожалуйста.',
+  'Просто продолжайте улыбаться. Когда-нибудь, миру надоест беспокоить вас.',
+  'Я просто прямой потомок грандиозности.',
+  'Историю составляют только люди, нарушающие правила.',
+  'Когда ты найдешь себя, жизнь меняется.',
+  'Всякий раз, когда вы чувствуете себя одиноким, вам нужно еще больше времени проводить самостоятельно.',
+  'Различие не всегда означает быть неправильным.',
+  'Вы можете судить меня только если вы совершенны.',
+  'Оставьте немного искорок, куда бы вы ни пошли.',
+  'Жизнь всегда кажется лучше, когда вы смеетесь.',
+  'Мир начинается с улыбки.',
+  'Я думаю, что настало время снова начать радоваться.',
+  'Я не совершенен. Я никогда им не был, и я никогда им не буду.',
+  'Прекратите пытаться исправить меня, потому что я не сломалась.',
+  'Мой успех просто отложен на будущее.',
+  'Все, что мне нужно, чтобы быть счастливой, — это шесть месяцев отпуска… Дважды в год.',
+  'Мне нравится быть мечтателем и ночным мыслителем.',
+  'Ничего не произойдет, пока вы не прекратите попытки.',
+  'Ваша скорость не имеет значения, пока вы продолжаете двигаться вперед.',
+  'Не должно быть никаких ярлыков, если вы решили оторваться!',
+  'Чтобы наслаждаться красивой радугой, вы должны сначала насладиться дождем.',
+  'Спокойно! Просто сегодня я сильно взволнована!',
+  'Улыбка — это самое короткое расстояние, которое вы можете проложить между двумя людьми.',
+  'Вам нужно просто быть немножко сумасшедшим, ведь маленькое сумасшествие очень похоже на небеса.',
+  'Просто убедитесь, что всегда влюблены в свою жизнь.',
+  'Просто продолжайте улыбаться — это сбивает с толку людей.',
+  'Никто не может потерпеть неудачу, пока им нравится своя жизнь.',
+  'Всегда следуйте своим надеждам, и не обращайте внимание на свои страхи.',
+  'Иногда лучшая месть — это улыбаться и двигаться дальше.',
+  'Улыбайся сейчас, потому что завтра может быть еще хуже.',
+  'Любовь и жизнь всегда найдут способ полюбить вас. Если конечно вы сами этого хотите.',
+  'Это просто доказательство того, что я могу быть лучшим эгоистом, чем вы.',
+  'Жизнь — это не вопрос достижения каждой цели, это празднование каждого момента.',
+  'Нельзя просто проживать жизнь. Ее следует праздновать.',
+  'Танцуйте, как будто никто не видит, любите, как никогда не любили. Иначе вы так и не узнаете, что такое жить по-настоящему.',
+  'Улыбка — наше последнее слово в марафете.',
+  'Как быть счастливой? Начинайте и заканчивайте каждый день с улыбкой.',
+  'Я особенный. Я уникален. Я — это я.',
+  'Я не ленива. Я просто очень расслаблена.',
+  'Улыбка может ответить на любой вопрос.',
+  'Я не такая, как они или кто-то еще. Я такая, как я.',
+  'Мой стиль уникален, поэтому пожалуйста, не копируйте его.',
+  'Жизнь слишком коротка, поэтому всегда ешь свой десерт в первую очередь.',
+  'Просто болтаюсь с самыми удивительными друзьями на свете.',
+  'Я знаю, что мне действительно повезло быть такой милахой.',
+  'Упади семь раз, но на восьмой — поднимайся.',
+  'Если вы чувствуете себя уверенно, тогда вы прекрасны.',
+  'Мой босс сказал мне, что у меня должен быть хоть один хороший день, поэтому я пошла домой.',
+  'Я люблю свою работу, особенно тогда, когда наступает отпуск.',
+  'Никакая красота не светит ярче, чем самое доброе сердце.',
+  'Вы можете улыбаться! Потому что это, наконец, пятница!',
+  'Никто не замечает, что я делаю, пока не забуду это сделать.',
+  'Моя мама сказала что нужно жить и мечтать, жить и мечтать. Поэтому я снова легла спать.',
+  'День за днем, планы на завтра… Но вечеринка сегодня!',
+  'Живите сегодня. Живите сейчас.',
+  'Будьте терпеливы, потому что лучшие вещи в жизни всегда требуют времени.',
+  'Вы можете себе представить? Я проснулась вот так!',
+  'Залог красоты — тренироваться каждый день так, как вы никогда раньше не тренировались.',
+  'Никогда не говорите людям о своей мечте. Вместо этого покажите им ее!',
+  'Поднимите голову, принцесса. Иначе Ваша корона свалится.',
+  'Моя жизнь похожа на открытую книгу, но не всем позволено читать оттуда.',
+  'Никогда не оглядывайся, когда любишь.',
+  'Сегодня я решила стать лучшей версией себя.',
+  'Улыбка — красивая кривая, которая делает мир прочнее.',
+  'Малыш, ты не справишься со мной, даже если я дам тебе инструкцию.',
+  'Я не боюсь шторма и урагана, потому что всегда плаваю со своим капитаном.',
+  'Мне нравится чувство уверенности в том, что я — это я.',
+  'Жизнь может быть не идеальной, но только не ваша одежда.',
+  'Несколько дней всегда начинаются лучше других.',
+  'Я не маленькая. У меня просто очень милый рост.',
+  'Всякий раз, когда у меня возникает проблема, я просто начинаю петь. Затем я понимаю, что мой голос — худшая из всех моих проблем.',
+  'Ненависть никого и никогда не сделала красивой.',
+  'Жизнь слишком коротка. Поэтому не забывайте улыбаться, пока ваши зубы все еще на месте.',
+  'Начинайте каждое утро с большущей улыбки.',
+  'Просто наслаждаюсь чашкой позитив-чая.',
+  'Работать трудно, но помните, что и хорошо погулять не просто.',
+  'Всегда окружайте себя людьми, которые делают вас счастливыми.',
+  'Всегда живи так, как ты мечтаешь.',
+  'Я заслуживаю того, чтобы это по-настоящему случилось.',
+  'Живите ради того, что вы любите больше всего, а не просто для того, чтобы произвести впечатление на других.',
+  'Я не ленива. Я просто включила режим энергосбережения.',
+  'Я не королева красоты, я просто красивая.',
+  'Я влюблена в себя всем своим сердцем.',
+  'Ни секунды не ждите, чтобы сделать себя счастливой.',
+  'И тогда я просто решила снова стать счастливой.',
+  'Подождите! Но сначала позвольте мне взять себя в руки.',
+  'Чем ярче светит солнце, тем меньше шансов увидеть тени жизни.',
+  'Улыбаться всегда легче, чем объяснять, почему ты грустишь.',
+  'Жизнь — это самая приятная коллекция счастливых моментов.',
+  'Никогда не недооценивайте скрытую силу искренней улыбки.',
+  'Улыбка — это свет, который светит через окно вашей души.',
+  'Наибольшие достижения всегда кажутся невозможными до тех пор, пока они не будут достигнуты.',
+  'Не нужно думать позитивно, нужно позитивно жить.',
+  'Время драгоценно, поэтому всегда тратьте его мудро.',
+  'Всегда держите голову высоко.',
+  'Я просто простой человек, который скрывает миллион чувств за одной счастливой улыбкой.',
+  'Солнце всегда светит в одиночестве, но его свет затмевает все.',
+  'Возьми мой совет, потому что я его не использую.',
+  'Жизнь состоит из маленьких моментов, именно так.',
+  'Мечты работают, если вы этого не сделаете.',
+  'Если возможности к вам не стучатся, тогда вам нужно просто построить дверь.',
+  'Жизнь продолжается. С тобой или без тебя.',
+  'Я сошла с ума. Вернитесь ко мне минут через пять.',
+  'Существует такая особая красота в том, чтобы просто быть собой.',
+  'Я бы хотела на несколько дней вернуться в прошлое. Нет, не исправлять ошибки, а лишь ощутить кое-что дважды.',
+  'Я заранее извиняюсь за все, что говорю, когда я голоден.',
+  'Имейте больше, чем вы когда-либо показывали, и говорите намного меньше, чем вы на самом деле знаете.',
+  'Улыбка — самая красивая вещь, которую вы можете носить.',
+  'Если ты по-настоящему счастлив, тебе все равно что думают об этом люди.',
+  'Все что имеет значение — просто быть собой.',
+  'Не пытайтесь учиться у меня, потому что это бессмысленно.',
+  'Я больше, чем вы видите или чего вы ожидаете.',
+  'Возможно, я не совершенна, но я определенно лимитированный экземпляр.',
+  'Я только встретила тебя, а оказалось, что это целое безумие.',
+  'Самое крутое доказательство того, что я умею делать фотки, лучше чем вы.',
+  'Всегда старайтесь быть счастливыми и делать счастливыми других.',
+  'Чтобы достичь новых берегов, мы должны плыть. Плыть, а не дрейфовать.',
+  'Котики, ну куда же без них?',
+  'С вами я забываю о всех своих проблемах.',
+  'Улыбнуться ничего не стоит, но каждая улыбка поистине бесценна.',
+  'В своей жизни ты должна играть только главную роль.',
+  'Я могу не победить сразу, но я обязательно добьюсь успеха.',
+  'Я люблю тебя. Это все, что я знаю.',
+  'Она — мой лучший друг.',
+  'Если я когда-то напишу историю нашей жизни, не удивляйся, что твое имя я упомяну несколько миллиардов раз.',
+];
+
+const createPhoto = () => ({
+  id: getId(),
+  url: getUrl(),
+  description: getRandomArrayElem(description),
+  likes: getRandomInt(MIN_LIKES, MAX_LIKES),
+  comments: getRandomInt(MIN_COMMENTS, MAX_COMMENTS),
+});
+
+const createPhotoGallery = () => Array.from({length: MAX_ID}, createPhoto);
+
+export {createPhotoGallery};
